Checking in the way fully implemented
The pieces are now capable of checking to see if everything is fully in the way, now we would just have to add something that sees if the king is being exposed. After that we'll just slap 'sees', 'checkPathOpen', and 'checkExposeKing' into another method and validate move will be completely done.