Lots of progress on validate move
The controller now has a (possibly finished) method to move pieces, and it includes individual validate moves as well as the method to check if the king will be put in check (and undo the move if that is the case). The part that sees if the king gets put in check needs testing still.